requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
langchain>=0.0.334
openai>=1.2.3
chromadb>=0.4.15
sentence-transformers>=2.2.2
tiktoken>=0.5.1
pdfminer.six>=20221105
python-docx>=1.0.1
groq>=0.5.0
cachetools>=5.3.0
